Revert "fixup: pull find::convert_all_modifiers() into new class in private/"
This reverts commit 1e6671f0e081cb4b232db645763af4dd9f834476.
